requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
deepl==1.18.0
idna==3.7
polib==1.2.0
python-dotenv==1.0.1
requests==2.31.0
setuptools==69.5.1
urllib3==2.2.1
